// packages/app/libs/constants/event-types-enum.ts
enum EventTypes {
  ABERTURA = "Abertura",
  PALESTRA = "Palestra",
  MINICURSO = "Minicurso",
  RODA = "Roda",
  FEIRA = "Feira",
  GAME_NIGHT = "Game Night",
  HACKATHON = "Hackathon",
  CONCURSO = "Concurso",
  CONTEST = "Contest",
  CULTURAL = "Cultural",
  COFFEE = "Coffee",
}

export default EventTypes;